code-model-finetuning/distil_glm5/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "distil-glm5"
version = "0.1.0"
description = "Scripts-first pipeline for GLM-5→GLM-4 synthetic distillation dataset extraction."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "pyyaml>=6.0",
  "requests>=2.31",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.9.0",
]
hf = [
  "datasets>=2.16",
]
dedup = [
  "datasketch>=1.6.5",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

